runq-disk: attach embedded disk files via virtio-blk
An embedded disk is a raw disk file located in the container filesystem
that can be attached to a VM via virtio-blk. The disk file will be created
if it does not already exist. It will also be resized if needed. Supported file
systems are ext2 and ext4. By default the disk file is created in the
<mount-dir> and therefore invisible inside the VM.

Embedded disks must be specified via the environment variable RUNQ_DISK.

Syntax of RUNQ_DISK:
  RUNQ_DISK="dir=<mount-dir>[,cache=<cache-type>,fstype=<filesystem-type>,id=<id>,size=<size>,label=<label>,mount=<yes|no>];..."

This simple example will create an embedded disk of 1GB mounted to /data:
  docker run --runtime runq -e RUNQ_DISK="dir=/data,size=1G" ...

Note: The deprecated syntax for external disks ("<source>:/dev/disk/<cache>/<filesystem>/<mountpoint>")
is no longer supported.

Signed-off-by: Peter Morjan <peter.morjan@de.ibm.com>