enable block device to be used as rootdisk
A block device can be used as rootdisk via virtio-blk. At first
boot of the container the content of the Docker image will be
coppied into the rootdisk. The rootdisk will then be used as rootfs
for the VM instead of 9pfs.
The block device must have an empty ext2 or ext4 filesystem.

Example:
  docker run \
    --runtime runq \
    --device /dev/sdc1:/dev/runq/0001/none/ext2 \
    -e RUNQ_ROOTDISK=0001 \
    -ti alpine sh

Directories can be excluded from being coppied:
 -e RUNQ_ROOTDISK_EXCLUDE="/foo,bar"

Signed-off-by: Peter Morjan <peter.morjan@de.ibm.com>