fix for using a restricted working directory
If the working directory is owned by a non root user with no exec
permissions for root e.g.:
     drwx------ 0700 1000  1000 /work
then
    docker run --workdir /work ...
fails with "chdir /work: permission denied".

The reason is that in this case syscall.Chdir() requires CAP_DAC_READ_SEARCH
capabilities. The fix is to execute chdir before dropping capabilities.

Signed-off-by: Peter Morjan <peter.morjan@de.ibm.com>