add support for the "clone3" syscall
Starting with Docker v20.10.10 the default seccomp profile includes
the new syscall "clone3". https://github.com/moby/moby/commit/9f6b562d

In order to support the fallback to "clone" in latest glibc 2.34 the errno
of the filter action SCMP_ACT_ERRNO must be set to "ENOSYS" instead of the
default "EPERM". The difficulties and challenges are well explained in
https://github.com/moby/moby/commit/9f6b562d

Signed-off-by: Peter Morjan <peter.morjan@de.ibm.com>