improve running container with Systemd
When stopping or restarting a container the signal sent form Docker to the
docker-entry point (runq-proxy) is SIGTERM. runq-proxy forwards it to runq-init
inside the VM and runq-int sends it to the "user" entry-point, e.g. Systemd.
But Systemd usually ignores SIGTERM. In order to trigger a clean Systemd shutdown
on "docker stop" the signal sent to Systemd must be changed to SIGRTMIN+4.
When Systemd receives SIGRTMIN+4 it starts the poweroff.target unit.
Runq now makes this change if the environment variable RUNQ_SYSTEMD is set to 1.

Signed-off-by: Peter Morjan <peter.morjan@de.ibm.com>